feat: EES-41 Final: added data fetching
