+++ src/utils/data/globalData.ts
@@ -1,0 +1,11 @@
+export enum Style {
+    placeHolder = "#6d6d6d",
+    mainColor = "#317334",
+    headerColor = "#30804b",
+  }
+  
+export enum loginErr {
+    error = "error",
+    success = "success",
+  }
+
